@@ -1,17 +1,17 @@
 apply plugin: 'com.android.application'
 
-def AAVersion = '4.3.1'
-def PSVersion = '15.0.1'
+def AAVersion = '4.6.0'
+def PSVersion = '16.0.0'
 def DBFlowVersion = '3.1.1'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.2"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "org.pb.android.geomap3d"
-        minSdkVersion 21
-        targetSdkVersion 26
+        minSdkVersion 23
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
         //testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -35,16 +35,17 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
     annotationProcessor "org.androidannotations:androidannotations:$AAVersion"
-    compile "org.androidannotations:androidannotations-api:$AAVersion"
+    implementation "org.androidannotations:androidannotations-api:$AAVersion"
 
-    compile 'org.greenrobot:eventbus:3.0.0'
+    implementation 'org.greenrobot:eventbus:3.1.1'
 
-    implementation 'com.android.support:appcompat-v7:26.1.0'
+    //implementation 'com.android.support:appcompat-v7:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation "com.google.android.gms:play-services-location:$PSVersion"
 
     annotationProcessor "com.github.Raizlabs.DBFlow:dbflow-processor:${DBFlowVersion}"
-    compile "com.github.Raizlabs.DBFlow:dbflow-core:${DBFlowVersion}"
-    compile "com.github.Raizlabs.DBFlow:dbflow:${DBFlowVersion}"
+    implementation "com.github.Raizlabs.DBFlow:dbflow-core:${DBFlowVersion}"
+    implementation "com.github.Raizlabs.DBFlow:dbflow:${DBFlowVersion}"
 
     //implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     //testImplementation 'junit:junit:4.12'
