@@ -3,11 +3,12 @@ package org.pb.android.geomap3d.location;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
 import android.util.Log;
 
 import org.androidannotations.annotations.Bean;
 import org.androidannotations.annotations.EService;
+
+import androidx.annotation.Nullable;
 
 @EService
 public class LocationService extends Service {
